rutabaga_gfx: Move RutabagaResource to separate module
Move RutabagaResource, Rutabaga2DInfo, and related snapshot structs from rutabaga_core into a new src/resource.rs module.
