+++ src/cross_domain/context.rs
@@ -69,8 +69,8 @@ use crate::cross_domain::cross_domain_protocol::CROSS_DOMAIN_MAX_IDENTIFIERS;
 use crate::cross_domain::cross_domain_protocol::CROSS_DOMAIN_QUERY_RING;
 use crate::cross_domain::worker::CrossDomainWorker;
 use crate::handle::RutabagaHandle;
+use crate::resource::RutabagaResource;
 use crate::rutabaga_core::RutabagaContext;
-use crate::rutabaga_core::RutabagaResource;
 use crate::rutabaga_core::VirtioFsLookup;
 use crate::rutabaga_utils::Resource3DInfo;
 use crate::rutabaga_utils::ResourceCreateBlob;
