+++ src/gfxstream.rs
@@ -40,9 +40,9 @@ use crate::handle::RutabagaHandle;
 use crate::renderer_utils::ret_to_res;
 use crate::renderer_utils::RutabagaCookie;
 use crate::renderer_utils::VirglBox;
+use crate::resource::RutabagaResource;
 use crate::rutabaga_core::RutabagaComponent;
 use crate::rutabaga_core::RutabagaContext;
-use crate::rutabaga_core::RutabagaResource;
 use crate::rutabaga_utils::DeviceId;
 use crate::rutabaga_utils::GfxstreamFlags;
 use crate::rutabaga_utils::ResourceCreate3D;
