+++ src/magma/context.rs
@@ -11,8 +11,8 @@ use magma_gpu::util::Handle as MagmaGpuHandle;
 use crate::context_common::ContextResource;
 use crate::context_common::ContextResources;
 use crate::handle::RutabagaHandle;
+use crate::resource::RutabagaResource;
 use crate::rutabaga_core::RutabagaContext;
-use crate::rutabaga_core::RutabagaResource;
 use crate::rutabaga_utils::ResourceCreateBlob;
 use crate::rutabaga_utils::RutabagaComponentType;
 use crate::rutabaga_utils::RutabagaFence;
