+++ src/rutabaga_core.rs
@@ -75,118 +75,8 @@ pub trait VirtioFsLookup: Send + Sync {
 const RUTABAGA_DEFAULT_WIDTH: u32 = 1280;
 const RUTABAGA_DEFAULT_HEIGHT: u32 = 1024;
 
-/// Information required for 2D functionality.
-#[derive(Clone, Deserialize, Serialize)]
-pub struct Rutabaga2DInfo {
-    pub width: u32,
-    pub height: u32,
-    pub host_mem: Option<Vec<u8>>,
-    pub scanout_stride: Option<u32>,
-}
-
-#[derive(Clone, Deserialize, Serialize)]
-struct Rutabaga2DSnapshot {
-    width: u32,
-    height: u32,
-    // NOTE: `host_mem` is not preserved to avoid snapshot bloat.
-}
-
-/// A Rutabaga resource, supporting 2D and 3D rutabaga features.  Assumes a single-threaded library.
-pub struct RutabagaResource {
-    pub resource_id: u32,
-    pub handle: Option<Arc<RutabagaHandle>>,
-    pub blob: bool,
-    pub blob_mem: u32,
-    pub blob_flags: u32,
-    pub map_info: Option<u32>,
-    pub info_2d: Option<Rutabaga2DInfo>,
-    pub info_3d: Option<Resource3DInfo>,
-    pub vulkan_info: Option<VulkanInfo>,
-    pub backing_iovecs: Option<Vec<RutabagaIovec>>,
-    /// Bitmask of components that have already imported this resource
-    pub component_mask: u8,
-    pub size: u64,
-    pub mapping: Option<MemoryMapping>,
-}
-
-/// The preserved fields of `RutabagaResource` that are saved and loaded across snapshot and
-/// restore.
-#[derive(Deserialize, Serialize)]
-struct RutabagaResourceSnapshot {
-    resource_id: u32,
-    // NOTE: `RutabagaResource::handle` is not included here because OS handles will
-    // not be valid across snapshot and restore.  The caller of `Rutagaba::restore()`
-    // is expected to re-map resources (via `Rutabaga::map()` or `Rutabaga::export_blob()`)
-    // when restoring snapshots.
-    blob: bool,
-    blob_mem: u32,
-    blob_flags: u32,
-    map_info: Option<u32>,
-    info_2d: Option<Rutabaga2DSnapshot>,
-    info_3d: Option<Resource3DInfo>,
-    vulkan_info: Option<VulkanInfo>,
-    // NOTE: `RutabagaResource::backing_iovecs` isn't snapshotted because the
-    // pointers won't be valid at restore time, see the `Rutabaga::restore` doc.
-    // If the client doesn't attach new iovecs, the restored resource will
-    // behave as if they had been detached (instead of segfaulting on the stale
-    // iovec pointers).
-    component_mask: u8,
-    size: u64,
-    // NOTE: `RutabagaResource::mapping` is not included here because mapped resources
-    // generally will not be mapped to the same host virtual address across snapshot
-    // and restore. The caller of `Rutagaba::restore()` is expected to re-map resources
-    // (via `Rutabaga::map()`) when restoring snapshots.
-}
-
-impl TryFrom<&RutabagaResource> for RutabagaResourceSnapshot {
-    type Error = RutabagaError;
-    fn try_from(resource: &RutabagaResource) -> Result<Self, Self::Error> {
-        Ok(RutabagaResourceSnapshot {
-            resource_id: resource.resource_id,
-            blob: resource.blob,
-            blob_mem: resource.blob_mem,
-            blob_flags: resource.blob_flags,
-            map_info: resource.map_info,
-            info_2d: resource.info_2d.as_ref().map(|info| Rutabaga2DSnapshot {
-                width: info.width,
-                height: info.height,
-            }),
-            info_3d: resource.info_3d,
-            vulkan_info: resource.vulkan_info,
-            size: resource.size,
-            component_mask: resource.component_mask,
-        })
-    }
-}
-
-impl TryFrom<RutabagaResourceSnapshot> for RutabagaResource {
-    type Error = RutabagaError;
-    fn try_from(snapshot: RutabagaResourceSnapshot) -> Result<Self, Self::Error> {
-        Ok(RutabagaResource {
-            resource_id: snapshot.resource_id,
-            handle: None,
-            blob: snapshot.blob,
-            blob_mem: snapshot.blob_mem,
-            blob_flags: snapshot.blob_flags,
-            map_info: snapshot.map_info,
-            info_2d: snapshot.info_2d.map(|info| {
-                let size = u64::from(info.width * info.height * 4);
-                Rutabaga2DInfo {
-                    width: info.width,
-                    height: info.height,
-                    host_mem: Some(vec![0; usize::try_from(size).unwrap()]),
-                    scanout_stride: None,
-                }
-            }),
-            info_3d: snapshot.info_3d,
-            vulkan_info: snapshot.vulkan_info,
-            backing_iovecs: None,
-            size: snapshot.size,
-            component_mask: snapshot.component_mask,
-            mapping: None,
-        })
-    }
-}
+use crate::resource::RutabagaResource;
+use crate::resource::RutabagaResourceSnapshot;
 
 /// A RutabagaComponent is a building block of the Virtual Graphics Interface (VGI).  Each component
 /// on it's own is sufficient to virtualize graphics on many Google products.  These components wrap
